+++ oci-test/tests/oci_core_boot_volume/variables.tf
@@ -57,7 +57,7 @@ locals {
 resource "null_resource" "test_image" {
   depends_on = [oci_core_subnet.named_test_resource]
   provisioner "local-exec" {
-    command = "oci compute image list --compartment-id ${var.tenancy_ocid} --all --display-name ${var.image} --output json > ${local.imagePath}"
+    command = "oci compute image list --compartment-id ${var.tenancy_ocid} --all --output json > ${local.imagePath}"
   }
 }