+++ oci-test/tests/oci_core_boot_volume_replica/test-list-query.sql
@@ -1,3 +1,3 @@
 select id, display_name, boot_volume_id
 from oci.oci_core_boot_volume_replica
-where boot_volume_id = '{{ output.boot_volume_id.value }}' and lifecycle_state = 'AVAILABLE';
+where boot_volume_id = '{{ output.boot_volume_id.value }}';