+++ oci-test/tests/oci_identity_tag_default/test-get-query.sql
@@ -1,3 +1,0 @@
-select id, lifecycle_state
-from oci.oci_identity_tag_default
-where id = '{{ output.resource_id.value }}';