+++ oci-test/tests/oci_identity_tag_default/test-turbot-query.sql
@@ -1,3 +1,3 @@
 select title, tenant_id
 from oci.oci_identity_tag_default
-where id = '{{ output.resource_id.value }}';
+where tag_definition_id = '{{ output.tag_definition_id.value }}';