+++ oci-test/tests/oci_logging_log/test-not-found-query.sql
@@ -1,3 +1,3 @@
 select name, id
 from oci.oci_logging_log
-where name = '{{ output.display_name.value }}::dummy';
+where name = 'dummy';