+++ oci/table_oci_identity_authentication_policy.go
@@ -8,7 +8,6 @@ import (
 	"github.com/turbot/steampipe-plugin-sdk/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/plugin"
 	"github.com/turbot/steampipe-plugin-sdk/plugin/transform"
-	"github.com/turbot/go-kit/types"
 )
 
 //// TABLE DEFINITION
@@ -110,7 +109,7 @@ func listAuthenticationPolicy(ctx context.Context, d *plugin.QueryData, _ *plugi
 
 	// The OCID of the tenancy containing the compartment.
 	request := identity.GetAuthenticationPolicyRequest{
-		CompartmentId: types.String(compartment),
+		CompartmentId: &session.TenancyID,
 		RequestMetadata: common.RequestMetadata{
 			RetryPolicy: getDefaultRetryPolicy(),
 		},