+++ oci/table_oci_mysql_db_system.go
@@ -433,7 +433,7 @@ func buildMySQLDBSystemFilters(equalQuals plugin.KeyColumnEqualsQualMap) mysql.L
 		request.DbSystemId = types.String(equalQuals["id"].GetStringValue())
 	}
 	if equalQuals["is_analytics_cluster_attached"] != nil {
-		request.IsAnalyticsClusterAttached = types.Bool(equalQuals["display_name"].GetBoolValue())
+		request.IsAnalyticsClusterAttached = types.Bool(equalQuals["is_analytics_cluster_attached"].GetBoolValue())
 	}
 	if equalQuals["is_heat_wave_cluster_attached"] != nil {
 		request.IsHeatWaveClusterAttached = types.Bool(equalQuals["is_heat_wave_cluster_attached"].GetBoolValue())